Initial Commit
This is the CS department user management application I wrote in
college.  It was used to store account users of CS systems in a MySQL
database because OS X server couldn't export passwords and it also
didn't have a decent webapp for password management. It also includes a
simple script to add users to an OS X server from a data source.